New: Added ability to recover experiments
New Property: Added `transmitValue` to Experiment.  A Value between 0
and 1 indicates the percentage of the overall science value that can be
transmitted.  If less than 1, the remaining science must be recovered.
New Property: Added `maximumData` to Experiment for use in 'open ended'
experiments that don't have a defined amount of data needed for
completion.
New Property: Added `scienceValuePerData` to Experiment for use in 'open
ended' experiments
New Property: Added `canFailAtAnyTime` to Experiment.  When true a fail
condition can cause an experiment to fail even after all data has been
collected or analyazed.  At any time.
New Property: Added `maximumDataModifier` to Conditions to allow a
condition to movidy the maximum data cap with a multiplier
New Property: Added `maximumDataBonus` to Conditions to allow a
condition to add a fixed bonus to the maximum data setting.